Cursor naming fixed
